# docs/reference/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: OpenManager AI API (v1)
  version: 2026.02.28
  contact:
    name: OpenManager Team
    email: support@openmanager-ai.vercel.app
  description: |
    Contract-first API baseline for high-complexity endpoints.
    Scope excludes complete product surface; remaining routes are in the route catalog.
servers:
  - url: /
    description: Vercel API base
paths:
...
  /api/ai/supervisor/stream/v2:
    get:
      operationId: resumeStreamV2
      summary: Resume streaming response
      security:
        - sessionCookie: []
        - secureSessionCookie: []
        - guestSessionCookie: []
        - apiKeyAuth: []
        - testSecretAuth: []
      parameters:
        - in: query
          name: sessionId
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
            pattern: '^[A-Za-z0-9_-]{8,128}$'
        - in: query
          name: skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Resumed stream (SSE)
          headers:
            X-Resumed:
              schema:
                type: string
                example: 'true'
            X-Session-Id:
              schema:
                type: string
            X-Stream-Id:
              schema:
                type: string
            X-Skip-Chunks:
              schema:
                type: string
            X-Stream-Protocol:
              schema:
                type: string
                example: 'ui-message-stream'
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE payload chunks
        '204':
          description: No active stream
        '400':
          description: invalid sessionId/skip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
            X-RateLimit-Daily-Limit:
              $ref: '#/components/headers/X-RateLimit-Daily-Limit'
            X-RateLimit-Daily-Remaining:
              $ref: '#/components/headers/X-RateLimit-Daily-Remaining'
            X-RateLimit-Daily-Reset:
              $ref: '#/components/headers/X-RateLimit-Daily-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
    post:
      operationId: createStreamV2
      summary: Create resumable stream
      security:
        - sessionCookie: []
        - secureSessionCookie: []
        - guestSessionCookie: []
        - apiKeyAuth: []
        - testSecretAuth: []
      parameters:
        - in: header
          name: X-Device-Type
          required: false
          schema:
            type: string
            enum: [mobile, desktop]
          description: Optional device hint forwarded to Cloud Run. Invalid values default to `desktop`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupervisorStreamRequest'
      responses:
        '200':
          description: Stream started (SSE)
          headers:
            X-Session-Id:
              schema:
                type: string
            X-Stream-Id:
              schema:
                type: string
            X-Stream-Protocol:
              schema:
                type: string
                example: 'ui-message-stream'
            X-Resumable:
              schema:
                type: string
                example: 'true'
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Stream pipeline error
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE error event stream (text/event-stream)
        '503':
          description: Streaming unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
            X-RateLimit-Daily-Limit:
              $ref: '#/components/headers/X-RateLimit-Daily-Limit'
            X-RateLimit-Daily-Remaining:
              $ref: '#/components/headers/X-RateLimit-Daily-Remaining'
            X-RateLimit-Daily-Reset:
              $ref: '#/components/headers/X-RateLimit-Daily-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    testSecretAuth:
      type: apiKey
      in: header
      name: x-test-secret
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
    secureSessionCookie:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
    guestSessionCookie:
      type: apiKey
      in: cookie
      name: auth_session_id
  headers:
    X-RateLimit-Limit:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Remaining:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Reset:
      schema:
        type: string
        description: Epoch ms timestamp
    X-RateLimit-Daily-Limit:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Daily-Remaining:
      schema:
        type: integer
        minimum: 0
    X-RateLimit-Daily-Reset:
      schema:
        type: string
        description: Epoch ms timestamp
    Retry-After:
      schema:
        type: string
        description: Retry delay in seconds
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
        fallback:
          oneOf:
            - type: boolean
            - type: string
        isServerError:
          type: boolean
        notFound:
          type: boolean
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: string
            - type: string
    RateLimitError:
      type: object
      required: [error, message, retryAfter]
      properties:
        error:
          type: string
          example: Too Many Requests
        message:
          type: string
        retryAfter:
          type: integer
          minimum: 0
        dailyLimitExceeded:
          type: boolean
...
    SupervisorStreamRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required: [role]
            properties:
              role:
                type: string
                enum: [user, assistant, system]
              content:
                type: string
              parts:
                type: array
                items:
                  type: object
                  additionalProperties: true
              createdAt:
                oneOf:
                  - type: string
                  - type: string
                    format: date-time
        sessionId:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9_-]{8,128}$'
        enableWebSearch:
          type: boolean
        enableRAG:
          type: boolean